Moved BroadcastReceiver for CustomTabs in own class
