+++ app/src/main/res/layout/webview.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:orientation="vertical" android:layout_width="fill_parent"
+    android:layout_height="fill_parent">
     <com.github.dfa.diaspora_android.ui.ContextMenuWebView
         android:id="@+id/webView"
         android:layout_width="fill_parent"
